+++ flutter_google_places_sdk/lib/flutter_google_places_sdk.dart
@@ -8,9 +8,8 @@ export 'package:flutter_google_places_sdk_platform_interface/flutter_google_plac
 /// Client used to call methods on the native google places sdk
 class FlutterGooglePlacesSdk {
   /// Construct a FlutterGooglePlacesSdk using the specific api key and locale
-  FlutterGooglePlacesSdk(this._apiKey, {Locale? locale, bool useNewApi = false})
-    : this._locale = locale,
-      this._useNewApi = useNewApi;
+  FlutterGooglePlacesSdk(this._apiKey, {Locale? locale})
+    : this._locale = locale;
 
   /// "Powered by google" image that should be used when background is white
   static const AssetImage assetPoweredByGoogleOnWhite =
@@ -34,8 +33,6 @@ class FlutterGooglePlacesSdk {
   Locale? get locale => _locale;
 
   Locale? _locale;
-
-  bool _useNewApi;
 
   Future<void>? _lastMethodCall;
   Future<void>? _initialization;
@@ -69,12 +66,11 @@ class FlutterGooglePlacesSdk {
   }
 
   Future<void> _ensureInitialized() {
-    return _initialization ??=
-        platform.initialize(apiKey, locale: locale, useNewApi: _useNewApi)
-          ..catchError((dynamic err) {
-            log('FlutterGooglePlacesSdk::_ensureInitialized error: $err');
-            _initialization = null;
-          });
+    return _initialization ??= platform.initialize(apiKey, locale: locale)
+      ..catchError((dynamic err) {
+        log('FlutterGooglePlacesSdk::_ensureInitialized error: $err');
+        _initialization = null;
+      });
   }
 
   /// Fetches autocomplete predictions based on a query.
@@ -233,21 +229,12 @@ class FlutterGooglePlacesSdk {
   /// Updates the settings of the places client with the given API key and locale.
   /// If apiKey is null, the last key will be used.
   /// If locale is null, it will not be updated.
-  Future<void> updateSettings({
-    String? apiKey,
-    Locale? locale,
-    bool? useNewApi,
-  }) {
+  Future<void> updateSettings({String? apiKey, Locale? locale}) {
     _apiKey = apiKey ?? this.apiKey;
     _locale = locale;
-    _useNewApi = useNewApi ?? _useNewApi;
 
     return _addMethodCall(
-      () => platform.updateSettings(
-        _apiKey,
-        locale: locale,
-        useNewApi: _useNewApi,
-      ),
+      () => platform.updateSettings(_apiKey, locale: locale),
     );
   }
 }