+++ flutter_google_places_sdk_http/lib/flutter_google_places_sdk_http.dart
@@ -38,11 +38,7 @@ class FlutterGooglePlacesSdkHttpPlugin
   }
 
   @override
-  Future<void> initialize(
-    String apiKey, {
-    Locale? locale,
-    bool useNewApi = false,
-  }) async {
+  Future<void> initialize(String apiKey, {Locale? locale}) async {
     _apiKey = apiKey;
     _locale = locale;
   }
@@ -51,11 +47,7 @@ class FlutterGooglePlacesSdkHttpPlugin
   Future<bool?> isInitialized() async => _apiKey != null;
 
   @override
-  Future<void> updateSettings(
-    String apiKey, {
-    Locale? locale,
-    bool? useNewApi,
-  }) async {
+  Future<void> updateSettings(String apiKey, {Locale? locale}) async {
     _apiKey = apiKey;
     if (locale != null) {
       _locale = locale;