+++ flutter_google_places_sdk_http/test/flutter_google_places_sdk_http_test.dart
@@ -18,7 +18,7 @@ void main() {
     String body, {
     int statusCode = 200,
     void Function(http.Request)? onRequest,
-  }) {
+  }) {  
     return MockClient((request) async {
       onRequest?.call(request);
       return http.Response(body, statusCode);